[SD-919] Address comments brought up during review
