+++ core/src/main/scala/quasar/physical/mongodb/collection.scala
@@ -46,7 +46,7 @@ object Collection {
       case Nil => \/-(clusterF)
       case first :: rest => for {
         db       <- DatabaseNameParser(first)
-        collSegs = rest.map(CollectionSegmentParser(_))
+        collSegs <- rest.traverseU(CollectionSegmentParser(_))
         coll     =  collSegs.mkString(".")
         _        <- if (utf8length(db) + 1 + utf8length(coll) > 120)
                       -\/(InvalidPathError("database/collection name too long (> 120 bytes): " + db + "." + coll))
@@ -126,9 +126,12 @@ object Collection {
 
     def char: Parser[String] = substitute(CollectionNameEscapes) | "(?s).".r
 
-    def apply(input: String): String = parseAll(seg, input) match {
-      case Success(seg, _) => seg
-      case failure : NoSuccess => scala.sys.error("doesn't happen")
+    /**
+      * @return If implemented correctly, should always return a [[String]] in the right hand of the [[Disjunction]]
+      */
+    def apply(input: String): PathError \/ String = parseAll(seg, input) match {
+      case Success(seg, _) => \/-(seg)
+      case failure : NoSuccess => -\/(InvalidPathError("failed to parse ‘" + input + "’: " + failure.msg))
     }
   }
