+++ core/src/main/scala/quasar/physical/mongodb/filesystem.scala
@@ -18,19 +18,13 @@ package quasar.physical.mongodb
 
 import quasar.Predef._
 import quasar.fp._
-import quasar._, Backend._, Errors._, Evaluator._
+import quasar._, Backend._, Errors._
 import quasar.fs._, Path._
 
 import scalaz._, Scalaz._
 import scalaz.concurrent._
 import scalaz.stream._
 import scalaz.stream.io._
-
-import com.mongodb.{WriteError => _, _}
-import com.mongodb.client._
-import com.mongodb.client.model._
-import org.bson._
-import scala.collection.JavaConverters._
 
 trait MongoDbFileSystem extends PlannerBackend[Workflow.Crystallized] {
   protected def server: MongoWrapper
@@ -164,106 +158,4 @@ object RenameSemantics {
     case Backend.Overwrite => Overwrite
     case Backend.FailIfExists => FailIfExists
   }
-}
-
-
-sealed trait MongoWrapper {
-  protected def client: MongoClient
-
-  import scala.collection.Iterator
-  implicit def toScalaIterator[A](cursor: MongoCursor[A]): Iterator[A] = {
-    new Iterator[A] {
-      def hasNext = cursor.hasNext
-      def next = cursor.next()
-    }
-  }
-
-  def genTempName(col: Collection): Task[Collection] = for {
-    start <- SequenceNameGenerator.startUnique
-  } yield SequenceNameGenerator.Gen.generateTempName(col.databaseName).eval(start)
-
-  private val db = Memo.mutableHashMapMemo[String, MongoDatabase] { (name: String) =>
-    client.getDatabase(name)
-  }
-
-  def exists(path: Collection): Task[Boolean] = Task.delay {
-    val databaseExists = client.listDatabaseNames().iterator().contains(path.databaseName)
-    if (databaseExists)
-      client.getDatabase(path.databaseName).listCollectionNames().iterator().contains(path.collectionName)
-    else false
-  }
-
-  // Note: this exposes the Java obj, so should be made private at some point
-  def get(col: Collection): Task[MongoCollection[Document]] =
-    Task.delay(db(col.databaseName).getCollection(col.collectionName))
-
-  def rename(src: Collection, dst: Collection, semantics: RenameSemantics): Task[Unit] = {
-    val drop = semantics match {
-      case RenameSemantics.Overwrite => true
-      case RenameSemantics.FailIfExists => false
-    }
-
-    if (src.equals(dst))
-      Task.now(())
-    else
-      for {
-        s <- get(src)
-        _ <- Task.delay(s.renameCollection(
-          new MongoNamespace(dst.databaseName, dst.collectionName),
-          (new RenameCollectionOptions).dropTarget(drop)))
-      } yield ()
-  }
-
-  def drop(col: Collection): Task[Unit] = for {
-    c <- get(col)
-    _ = c.drop
-  } yield ()
-
-  def dropDatabase(name: String): Task[Unit] = Task.delay(db(name).drop)
-
-  def dropAllDatabases: Task[Unit] =
-    databaseNames.flatMap(_.traverse_(dropDatabase))
-
-  def insert(col: Collection, data: Vector[Document]): Task[Unit] = for {
-    c <- get(col)
-    _ = c.bulkWrite(data.map(new InsertOneModel(_)).asJava)
-  } yield ()
-
-  def databaseNames: Task[Set[String]] =
-    MongoWrapper.databaseNames(client)
-
-  def collections: Task[List[Collection]] =
-    databaseNames.flatMap(_.toList.traverse(MongoWrapper.collections(_, client)))
-      .map(_.join)
-}
-
-object MongoWrapper {
-  def apply(client0: com.mongodb.MongoClient) =
-    new MongoWrapper { val client = client0 }
-
-  def liftTask: (Task ~> EvaluationTask) =
-    new (Task ~> EvaluationTask) {
-      def apply[A](t: Task[A]) =
-        EitherT(t.attempt).leftMap(e => CommandFailed(e.getMessage))
-    }
-
-  /**
-   Defer an action to be performed against MongoDB, capturing exceptions
-   in EvaluationError.
-   */
-  def delay[A](a: => A): EvaluationTask[A] =
-    liftTask(Task.delay(a))
-
-  def databaseNames(client: MongoClient): Task[Set[String]] =
-    Task.delay(try {
-      client.listDatabaseNames.asScala.toSet
-    } catch {
-      case _: MongoCommandException =>
-        client.getCredentialsList.asScala.map(_.getSource).toSet
-    })
-
-  def collections(dbName: String, client: MongoClient): Task[List[Collection]] =
-    Task.delay(client.getDatabase(dbName)
-      .listCollectionNames.asScala.toList
-      .map(Collection(dbName, _)))
 }
